extract Mcrain::ContainerController
